+++ shell/crates/running_apps/src/main.rs
@@ -1,8 +1,8 @@
 use commons::prelude::*;
 use gpui::prelude::*;
 use gpui::*;
-use running_apps::models::models::{ AppCard, AppMessage, RunningApps };
-use running_apps::prelude::app_manager::AppManagerService;
+use running_apps::models::models::{ AppCard, RunningApps };
+use running_apps::prelude::app_manager::{ AppManagerService, AppMessage };
 use tokio::sync::mpsc;
 
 fn main() {
