+++ shell/crates/running_apps/src/services/mod.rs
@@ -1,3 +1,2 @@
    pub mod app_manager;
    pub mod desktop_entries;
-   pub mod desktop_models;
