+++ shell/crates/running_apps/src/ui/components/card.rs
@@ -1,7 +1,8 @@
 use gpui::prelude::*;
 use gpui::*;
 use crate::config::constants::*;
-use crate::models::models::{ AppManagerMessage, DragDirection };
+use crate::models::models::{  DragDirection };
+use crate::prelude::app_manager::AppManagerMessage;
 use crate::ui::{ CardDragData, RunningApps };
 use crate::ui::icon::{ Icon, IconName };
